requirements.txt
asgiref==3.3.1

cmapy==0.6.6
colormath==3.0.0

Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0

mysqlclient==2.0.3

PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2021.1

sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0

